requirements.txt
paho-mqtt
protobuf
setuptools
shortuuid
keyboard
cryptography